Exclude the ones with properties. Using regex instead of introspection for now.
