+++ src/Annotator/CommandAnnotator.php
@@ -2,9 +2,12 @@
 
 namespace IdeHelper\Annotator;
 
+use IdeHelper\Annotator\Traits\ModelTrait;
 use RuntimeException;
 
 class CommandAnnotator extends AbstractAnnotator {
+
+	use ModelTrait;
 
 	/**
 	 * @param string $path Path to file.
@@ -46,24 +49,6 @@ class CommandAnnotator extends AbstractAnnotator {
 		$modelName = $matches[1];
 
 		return $modelName;
-	}
-
-	/**
-	 * @param string $content
-	 *
-	 * @return array<string>
-	 */
-	protected function getUsedModels(string $content): array {
-		preg_match_all('/\$this->([a-z]+)\s*=\s*\$this->fetchTable\(\'([a-z.]+)\'/i', $content, $matches);
-		if (empty($matches[1])) {
-			return [];
-		}
-
-		$properties = $matches[1];
-		$tables = $matches[2];
-		$models = array_combine($properties, $tables);
-
-		return $models;
 	}
 
 }
