+++ tests/test_app/src/Command/MyCommand.php
@@ -7,11 +7,15 @@ class MyCommand extends Command {
 
 	protected ?string $defaultTable = 'Cars';
 
+	/** @var \Relations\Model\Table\BarsTable */
+	protected $Bars;
+
 	/**
 	 * @return void
 	 */
 	public function main() {
 		$this->Wheels = $this->fetchTable('Wheels');
+		$this->Bars = $this->fetchTable('Relations.Bars');
 	}
 
 }
