+++ tests/test_files/Command/MyCommand.php
@@ -11,11 +11,15 @@ class MyCommand extends Command {
 
 	protected ?string $defaultTable = 'Cars';
 
+	/** @var \Relations\Model\Table\BarsTable */
+	protected $Bars;
+
 	/**
 	 * @return void
 	 */
 	public function main() {
 		$this->Wheels = $this->fetchTable('Wheels');
+		$this->Bars = $this->fetchTable('Relations.Bars');
 	}
 
 }
